Completely reworked the databaseConnection hydrator
This addresses quite a few things:

When Mikro calls `toJSON` of an entity all relationships are lost and you get the value of a PK of that entity (which is also the value stored in the DB row). To address this we can automatically create a DTO-esque type called JSONified. It will convert any relations on a class definition to an int instead.

All places that used the post-toJSON version now take a JSONified type instead

The way the `References` were being created in the hydrator was just blatantly wrong. It's now correct and tested working.

A regression was fixed- Mikro doesn't allow for `upsert` calls on an entity with no PK defined. The databaseConnection's `upsert` has been updated to use the old method of `insert` + `findOne` when no PK is present (creating a new entity).