Update datastoreAbstract with get methods
This brings the datastoreAbstract back in line with the proper usage of a svelte store, in particular because the `get()` method is no longer a synchronous getter for retrieving a value that's already in the store. Instead, it triggers a hydrate for the given ID you're trying to get and returns a promise.
This also adds a `getWhere` method to coincide with the new method on the datastoreProvider interface
Lastly, `getAll` now triggers a hydrate