Fixed bug where newly created entities could not have many to many relations created during initial insert
Many to many realtionships use a linking table. In order to use this the ID for both linked entities must already exist. That's the purpose of the shallow clone here (and why m:n fields are ignored). The problem was that the old code never called em.assign after that insert to actually put in the relationship. (Updates did, but not inserts). Now, both inserts and updates persist the main entity data and then use the same em.assign() code to ensure all relationship changes are put in place.